Minor fixes

+++ src/components/RecurringPayments/Bills/BillDetails.js
@@ -112,16 +112,19 @@ const BillDetails = () => {
     }
 
     const handleDelete = async (billId) => {
-        setIsLoading(true);
-        try {
-            await deleteBillAsync(user.uid, billId);
-            dispatch(deleteBill(billId));
-        }
-        catch (ex) {
-            console.log(ex.message)
-        }
-        finally {
-            setIsLoading(false);
+        const response = window.confirm(`Confirm deletion`);
+        if (response) {
+            setIsLoading(true);
+            try {
+                await deleteBillAsync(user.uid, billId);
+                dispatch(deleteBill(billId));
+            }
+            catch (ex) {
+                console.log(ex.message)
+            }
+            finally {
+                setIsLoading(false);
+            }
         }
     }
